REFACTOR endpoint forms to one form and implement isLoading context state
