+++ pwa/src/context/isLoading.ts
@@ -6,6 +6,7 @@ export interface IsLoadingProps {
   authenticationForm?: boolean;
   sourceForm?: boolean;
   applicationForm?: boolean;
+  endpointForm?: boolean;
 }
 
 export const isLoading = {} as IsLoadingProps;
