+++ pwa/src/templates/templateParts/endpointsForm/EndpointsFormTemplate.module.css
@@ -1,18 +1,4 @@
-.container {
-  background-color: var(--gateway-ui-color-white);
-  padding-inline-start: var(--gateway-ui-size-md);
-  padding-inline-end: var(--gateway-ui-size-md);
-  padding-block-start: var(--gateway-ui-size-md);
-  padding-block-end: var(--gateway-ui-size-md);
-}
-
 .gridContainer {
-  margin-block-end: var(--gateway-ui-size-2xl);
-}
-
-.section {
-  display: flex;
-  justify-content: space-between;
   margin-block-end: var(--gateway-ui-size-2xl);
 }
 
@@ -20,41 +6,8 @@
   margin-block-end: var(--gateway-ui-size-2xl);
 }
 
-.buttons {
-  display: flex;
-  align-items: center;
-  justify-content: flex-end;
-}
-
-.buttons > * {
-  display: flex;
-  align-items: center;
-}
-
-.buttons *:not(:last-child) {
-  margin-inline-end: var(--gateway-ui-size-xs);
-}
-
-.button {
-  width: max-content;
-}
-
-.buttonIcon > span > svg {
-  margin-inline-end: var(--gateway-ui-size-xs);
-}
-
 .grid {
   display: grid;
   grid-gap: var(--gateway-ui-size-lg);
   grid-template-columns: 1fr 1fr;
-}
-
-.deleteButton {
-  --denhaag-button-primary-action-background-color: var(--gateway-ui-color-red);
-}
-
-.deleteButton:hover {
-  --denhaag-button-primary-action-hover-background-color: var(
-    --gateway-ui-color-dark-red
-  );
 }
